Refactor shared source portability helpers
