+++ crates/shuck-linter/src/rules/portability/source_common.rs
@@ -1,0 +1,85 @@
+use shuck_ast::{Command, Redirect, SimpleCommand, Span};
+use shuck_semantic::ScopeKind;
+
+use crate::{Checker, ShellDialect};
+
+#[derive(Clone, Copy)]
+pub(super) enum SourceScopeFilter {
+    InsideFunction,
+    OutsideFunction,
+}
+
+impl SourceScopeFilter {
+    fn matches(self, inside_function: bool) -> bool {
+        match self {
+            Self::InsideFunction => inside_function,
+            Self::OutsideFunction => !inside_function,
+        }
+    }
+}
+
+pub(super) fn source_command_spans_in_sh(
+    checker: &Checker<'_>,
+    filter: SourceScopeFilter,
+) -> Vec<Span> {
+    if !matches!(checker.shell(), ShellDialect::Sh | ShellDialect::Dash) {
+        return Vec::new();
+    }
+
+    let source = checker.source();
+    checker
+        .facts()
+        .commands()
+        .iter()
+        .filter(|fact| fact.effective_name_is("source"))
+        .filter(|fact| filter.matches(inside_function(checker, fact.span())))
+        .map(|fact| source_anchor_span(fact.command(), fact.redirects(), fact.span(), source))
+        .collect()
+}
+
+fn inside_function(checker: &Checker<'_>, span: Span) -> bool {
+    let scope = checker.semantic().scope_at(span.start.offset);
+    checker
+        .semantic()
+        .ancestor_scopes(scope)
+        .any(|scope| matches!(checker.semantic().scope_kind(scope), ScopeKind::Function(_)))
+}
+
+fn source_anchor_span(
+    command: &Command,
+    redirects: &[Redirect],
+    fallback: Span,
+    source: &str,
+) -> Span {
+    match command {
+        Command::Simple(command) => {
+            clip_first_line(simple_command_span(command, redirects), source)
+        }
+        _ => clip_first_line(fallback, source),
+    }
+}
+
+fn simple_command_span(command: &SimpleCommand, redirects: &[Redirect]) -> Span {
+    let start = command
+        .assignments
+        .first()
+        .map_or(command.name.span.start, |assignment| assignment.span.start);
+    let mut end = command
+        .args
+        .last()
+        .map_or(command.name.span.end, |word| word.span.end);
+    if let Some(redirect) = redirects.last() {
+        end = redirect.span.end;
+    }
+    Span::from_positions(start, end)
+}
+
+fn clip_first_line(span: Span, source: &str) -> Span {
+    let text = span.slice(source);
+    let Some(line_end) = text.find('\n') else {
+        return span;
+    };
+
+    let first_line = text[..line_end].trim_end_matches('\r');
+    Span::from_positions(span.start, span.start.advanced_by(first_line))
+}
